pyproject.toml
[project]
name = "mcp-llms-txt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "markdownify>=1.1.0",
    "mcp[cli]>=1.4.1",
    "pyyaml>=6.0.1",
]

[project.scripts]
llms-txt = "mcp_llms_txt.cli:main"

[dependency-groups]
test = [
    "ruff>=0.11.0",
]
